+++ base_qry_sqlcalls.php
@@ -15,7 +15,7 @@
 //          Author(s): Nathan Gibbs
 //                     Kevin Johnson
 
-if ( isset($join_sql) ){ // Issue #5
+if ( isset($join_sql) || $printing_ag ){ // Issue #5
 	GLOBAL $colored_alerts, $debug_mode;
 	// Run the Query.
 	// base_ag_main.php will include this file
@@ -29,6 +29,7 @@ if ( isset($join_sql) ){ // Issue #5
 		$page = "base_qry_main.php";
 		$cnt_sql = "SELECT COUNT(acid_event.cid) FROM acid_event ".$join_sql.$where_sql.$criteria_sql;
 		$tmp_page_get = "";
+		$sql .= $join_sql.$where_sql.$criteria_sql;
 	}
 	// UI Language Vars
 	$CPSig = $UIL->CWA['Sig'];
@@ -68,39 +69,37 @@ if ( isset($join_sql) ){ // Issue #5
 	$qro->AddTitle($CPL4P,
 		"proto_a", " ", "$OB ip_proto ASC",
 		"proto_d", " ", "$OB ip_proto DESC"
-	);
-	if ( !$printing_ag ){
-		$sql = $sql.$join_sql.$where_sql.$criteria_sql;
-	}
-	// Apply sort criteria.
-	if ( $qs->isCannedQuery() ){
-		$sql = $sql."$OB timestamp DESC ";
+);
+	if ( $qs->isCannedQuery() ){ // Apply sort criteria.
+		$sort_sql = "$OB timestamp DESC ";
 	}else{
-     $sort_sql = $qro->GetSortSQL($qs->GetCurrentSort(), $qs->GetCurrentCannedQuerySort());
-     //  3/23/05 BDB   mods to make sort by work for Searches
-     $sort_sql = "";
+		$sort_sql = $qro->GetSortSQL($qs->GetCurrentSort(), $qs->GetCurrentCannedQuerySort());
+		if ( !is_null($sort_sql) ){ // Issue #168
+			$sort_sql = $sort_sql[1]; // Issue #133 fix.
+		}
 		if (!isset($sort_order)) {
 			$sort_order = NULL;
 		}
-		// Issue #133 fix.
-		$sort_sql = $qro->GetSortSQL($qs->GetCurrentSort(), '')[1];
-     ExportHTTPVar("prev_sort_order", $sort_order);
-    
-     $sql = $sql." ".$sort_sql;
-  }
-
-		if ( $debug_mode > 0 ){
-			print "<br/>SUBMIT: $submit <br/>";
-			print "sort_order: $sort_order <br/>";
-			print "SQL (save_sql): $sql <br/>";
-			print "SQL (sort_sql): $sort_sql <br/>";
+		ExportHTTPVar("prev_sort_order", $sort_order);
+	}
+	$sql .= $sort_sql;
+	if ( $debug_mode > 0 ){
+		$TK = array ( 'SUBMIT', 'sort_order', 'SQL (save_sql)', 'SQL (sort_sql)' );
+		$DI = array($submit, $sort_order, $sql, $sort_sql );
+		$DD = array();
+		foreach ( $TK as $val ){
+			array_push($DD, $val);
 		}
-
-  /* Run the Query again for the actual data (with the LIMIT) */
-  //$result = ""; // $qs->ExecuteOutputQuery($sql, $db);
-  $result = $qs->ExecuteOutputQuery($sql, $db);
-  $et->Mark("Retrieve Query Data");
-
+		if ( $printing_ag ){
+			$ttmp = 'Alert Group';
+		}else{
+			$ttmp = 'Query';
+		}
+		DDT($DI,$DD, "$ttmp Debug", '', '',1);
+	}
+	// Run the Query again for the actual data (with the LIMIT), if any.
+	$result = $qs->ExecuteOutputQuery($sql, $db);
+	$et->Mark("Retrieve Query Data");
 	if ( $debug_mode > 0 ){
 		if ( $qs->isCannedQuery() ){
 			$CCF = 'Yes';